reorganized and added unit tests
